Refactor doctype display label logic as mixin (#1366)
